Custom locations: refactoring
